+++ src/App.scss
@@ -1,6 +1,3 @@
 iframe {
   display: none;
 }
-.dropdown-item {
-  cursor: pointer;
-}
