+++ src/Autocomplete/Autocomplete.scss
@@ -1,0 +1,3 @@
+.dropdown-item {
+  cursor: pointer;
+}
